refactor: rename to limit config to reflect actual usage
